Backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
sse-starlette
openai>=1.40.0
google-generativeai>=0.8.0
sentence-transformers
qdrant-client
redis
numpy
pypdf
python-docx
docx2python
pymupdf
apscheduler
PyJWT
passlib[bcrypt]
